First version of the PWA client.
In this early version we set up all the needed machinery to serve the
Progressive WebApp client. Most of the URLs had to be changed to be put
within the scope of the app.

When this change lands in production the clients [1] will be able to add
Emotly on the home screen; the app will be kinda usable also when the
user is offline or with bad data connection.

A lot of field testing will be necessary to define all the corner cases.

[1] ATM the only platform to support (almost) all the features we make
use of is Android/Chrome.