WebApp Manifest served in /.
The WebApp Manifest was ignored since it landed in production; it turned
out that it had to be served from the root for the UA to consider it.

We're fixing this here by mocking the /manifest.json URL by serving
whatever is in static/app/manifest.json.

Probably this is a bug of the implementer as nothing like this has been
found in the spec [1]. We'll have to revisit this issue in the future,
therefore a relevant FIXME has been added.

[1] https://w3c.github.io/manifest/